requirements.txt
Flask
Flask-Cors
Flask-SocketIO
Flask-Session
flask_pymongo
eventlet==0.30.2
gunicorn
python-decouple
pytz